simple wallet connect wrapper with auth check and decoupled disconnect button
